# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJN Luxury Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: inStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /service-orders:
    post:
      operationId: createServiceOrder
      tags: [services]
      summary: Submit a service order request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceOrderInput"
      responses:
        "201":
          description: Service order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/track/{trackingCode}/respond:
    post:
      operationId: respondToBooking
      tags: [services]
      summary: Customer confirms or requests a reschedule for a booking (public, gated by tracking code)
      parameters:
        - name: trackingCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingResponseInput"
      responses:
        "200":
          description: Updated booking tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"

  # ─── CART ────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── DELIVERY ────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Customer token returned by /auth/verify-otp.
    AdminCookieAuth:
      type: apiKey
      in: cookie
      name: ajn_admin_session
      description: HttpOnly admin session cookie returned by /admin/auth/login.
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Product:
      type: object
      required: [id, name, nameAr, price, stock, images, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        stock:
          type: integer
        category:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, nameAr, price, stock]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        stock:
          type: integer
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        stock:
          type: integer
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean

    # ─── SERVICES ─────────────────────
    Service:
      type: object
      required: [id, name, nameAr, type, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        type:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
    ServiceOrderInput:
      type: object
      required: [serviceId, customerName, phone, eventDate]
      properties:
        serviceId:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        eventDate:
          type: string
        eventLocation:
          type: string
        notes:
          type: string
        customFields:
          type: object
    ServiceOrder:
      type: object
      required: [id, serviceId, customerName, phone, status, createdAt]
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        trackingCode:
          type: ["string", "null"]
        customerName:
          type: string
        phone:
          type: string
        eventDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string

    # ─── CART ─────────────────────────
...
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: string
        quantity:
          type: integer
        price:
          type: number
        selectedColor:
          type: ["string", "null"]
        customization:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
...
    OrderTracking:
      type: object
      required: [trackingCode, status, createdAt]
      properties:
        trackingCode:
          type: string
        status:
          type: string
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        kind:
          type: string
          description: "product | service"
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
        createdAt:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
        mapsUrl:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        customerConfirmation:
          type: ["string", "null"]
          description: "confirmed | reschedule_requested | null"
        requestedDate:
          type: ["string", "null"]
        confirmationNote:
          type: ["string", "null"]
        confirmationAt:
          type: ["string", "null"]
    BookingResponseInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [confirm, reschedule]
        requestedDate:
          type: string
        note:
          type: string
    StatusHistoryEntry:
      type: object
      required: [status, createdAt]
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── GALLERY ──────────────────────
...
    Review:
      type: object
      required: [id, productId, customerName, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [productId, customerName, rating]
      properties:
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string

    # ─── DELIVERY ─────────────────────
...