# lib/api-zod/src/generated/types/bookingResponseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingResponseInputAction } from "./bookingResponseInputAction";

export interface BookingResponseInput {
  action: BookingResponseInputAction;
  requestedDate?: string;
  note?: string;
}