# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  nameAr: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  type: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  image?: string | null;
  isActive: boolean;
}