# lib/api-zod/src/generated/types/serviceOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceOrderInputCustomFields } from "./serviceOrderInputCustomFields";

export interface ServiceOrderInput {
  serviceId: number;
  customerName: string;
  phone: string;
  eventDate: string;
  eventLocation?: string;
  notes?: string;
  customFields?: ServiceOrderInputCustomFields;
}